Add ability to use config file and report exit code if any commandline commands fail (#4)
Add an additional field, credentials, that gets passed to the --defaults-extra-file flag. Credentials should be a filepath to a .cnf file containing additional options you would like to pass to the mysql commandline. We are only using it for credentials at the moment, but it could be used for hosts and databases as well. Add visual feedback for data fetching and mute version checking.