Make creating and dropping databases safer with "if (not) exists" check
[Background]
MySQL errors out if you try to create a table that already exists
or drop a table that does not exist.

[Changes]
This adds two methods so that those operations can be attempted in a
safe way (only create a database if it doesn't already exist in the
case of creating, and only delete a database if it already exists
in the case of dropping).

[Result]
Safer database creation and deletion